removed copypaste parts from rails amdin
