+++ config/initializers/admin_tabs_override.rb
@@ -1,4 +1,4 @@
-module AdminModule
+RailsAdmin::ApplicationHelper.module_eval do
   def menu_for(parent, abstract_model = nil, object = nil, only_icon = false) # perf matters here (no action view trickery)
     actions = actions(parent, abstract_model, object).select { |a| a.http_methods.include?(:get) }
     actions.collect do |action|
@@ -13,5 +13,4 @@ module AdminModule
       )
     end.join.html_safe
   end
-
 end
